Move fault detection state into caller function
The original implementation for checking for grounded signal inputs uses
a function local static variable to store state between calls, which
prevented the reuse of that logic for detecting other errors that have
some time component. This commit extracts that state to a struct that
caller functions can make static, so that the common fault checking
logic can be reused for other cases.