fault detection: distinguish between faults and fault conditions
Some fault conditions have a time component where an error state has to
be present for longer than an acceptable time period. In order to
distinguish between these bad states and an actual fault detection this
uses `condition` to indicate the anomalous state and `fault` to indicate
a state that has been anomalous for longer than an acceptable period.