Add BOM for each board
Prior to this commit there was not a reliable BOM associated with the
boards we tracked schematics. This commit represents the addition of
the BOM for each board as well as a corresponding MacroFab XYRS data
file. Each of these files are in a comma delimited CSV format that
can be imported into spreadsheet software.

Closes #227