Update dbc to support float
Prior to this commit the dbc file did not support the float leaving the user to
determine how to interpret the 32 bits. The variable naming was also a mix of
upper and lower case letters. This commit fixes that by setting the
command messages to the standard IEEE float and making the message parameters
all lower case.