testing: Reorganize test setup files to match best practices
The module firmware `common.cpp` file layout was a bit mangled due to
the test refactoring, which was confusing and a bit jarring. This commit
cleans up the unit test setup files to use a more consistent layout and
renames them to indicate that they are, in fact, just for setup.