Remove can filtering in detection
Prior to this commit there was CAN filtering in the detection portion of the
code. For unknown reasons this caused intermittent behavior where zero CAN
messages would be returned yielding nothing being detected and requiring the
user to close and restart the program. This commit fixes that by removing the
filtering during detection and relying more heavily on the for loop to check
which CAN IDs are included in the messages.