Reduce wait time and total messages scanned
Prior to this commit the detection took a while causing the user to wait long
enough to wonder if the program hung. This commit fixes that by reducing the
can message timeout and number of CAN IDs looked at to something more
reasonable.

The new timeout was calculated based on 500k baudrate at standard
message sizes showing 6250 messages per second. The new CAN ID number was
determined by number of messages on the CAN bus which was ~40 at a 40% bus
load.