Remove CAN Filter and Fix Return Pointer
Prior to this commit the code had CAN filters which are intermittently filtering
out everything. Also the get_device_name function was returning a char* which
is not adviced in C since that is the function's memory space. This commit fixes
these issues by removing CAN filtering and passing in a pointer for the
get_device_name function to modify.