Clean up structs and fix string initialization
Prior to this commit the code had a struct typedef with unused naming and the
struct default value was a global variable which could be changed. The code also
had a missing variable initialization for a string copy. This commit fixes that
by moving the struct initiailzation to the variable decleraction and using calloc
rather than malloc for the string intialization.