Fix build indexing
Prior to this commit the same node that constructed the multiple builds also
called the build function before the array had been constructed causing the
variables to be unpopulated at build time. This commit fixes that by having a
worker node construct the array of parallel builds and then call the builds
after the list exists.