Update Jenkinsfile to use cmake variable
Prior to this commit the platform variable did not work because the strings were
encapsulated in single quotes rather than double. Also the platform variable was
populated inside the jenkins file rather than some other location where the
available vehicles to build are located. This commit fixes that by encapsulating
strings in double quotes so the variable is evaluated and uses cmake to
determine how many platforms to build for.