Update Jenkins Tests
Prior to this commit the unit tests was missing the TESTS flag and the property
tests were building their own cmake directory which was not nessisary. This
commit fixes that by adding the TESTS flag and using the same test directory for
both the unit tests and the property tests.