Fix cucumber port clash and property tests
Prior to this commit the unit tests failed on parallel builds due to port
number clash and the property tests build due to cmake directory build process
missing. This commit fixes these by setting a port based on the jenkins executor
and the platform index to ensure a unique port used per test.