Add NULL firmware implementation
Prior to this commit there was no clear firmware solution for disabling one or a subset of OSCC modules. This commit represents the addition of a NULL firmware implementation such that any combination of modules can be essentially ignored by OSCC. The NULL implementation is build by default and uploaded to any module with the command `make null-upload`.