Fix voltage spoof ranges
Prior to this commit some of the voltage spoof ranges did not calculate to the
correct value when multiplying and dividing by the STEPS_PER_VOLT against the
the x_SPOOF_y_SIGNAL_VOLTAGE_z values. This commit fixes that by setting the
x_SPOOF_y_SIGNAL_RANGE_z to something that calculates correctly.