Correct return code check for oscc_close()
In oscc_close() when we were checking if the file descriptors could be
closed the sign check was reversed, and when the socket(s) were closed
the return code assumed that `close()` would return >0 when it returns 0
on success and -1 on error. In addition the variable naming introduced
variable shadowing of the function return code; setting the result
inside of close would instead set the shadowing variable whose value
would be discarded.

This commit resolve these issues by
  1) correcting the sign check to close opened file descriptors
  2) correcting the return code check of close()
  3) removing the variable shadowing to properly return closed values.

Test coverage of this is defined in the HIL property tests.