Clarify return code behavior for oscc_close()
The previous commit cleaned up the return code generation for
`oscc_close()` but did not resolve some of the uncertain semantics
around the return code of that function.  This commit changes the return
code for `oscc_close()` to only return `OSCC_OK` if one or more channels
closed, and no errors were encountered while closing channels.