Remove Null default
Prior to this commit null had the default uno setting locally which would
build uno regardless of the higher level CMake Arduino default board. This
commit fixes that by removing the local one so that if the global setting is
changed the null implementation is built on the global setting.