Add "getter" functions to unpack OBD-II CAN data
Prior to this commit the expected method for getting relevant data
from the DriveKit OBD-II messages was to cast the CAN frame data as
a C struct. The C structs defined by the API did not reflect the
CAN data in most cases which has the potential to mislead users
and developers who want to use the unpacked CAN data.

This commit represents the addition of "getter" functions
that take as a parameter, the CAN frame read and the value to set.
Each functoin unpacks its respective value from the CAN frame reported
by the vehicle and scales it according to it's unit of measurement (degrees, bar, kph).

Additionally, comments indicating that the old C structs are deprecated were added to the
documentation as removing them completely would represent a breaking change.