Isolated duplicate code for wheel speed
Prior to this commit the logic to unpack the wheel speed reported by the
vehicle contained duplicate code. This resulted in the need to make one
change in 4 places and the potential for difficult to spot bugs. After
this commit the duplicate code has been isolated to its own function.