Add vehicle + OSCC consensus checking tool
Prior to this commit there was no easy way to verify that OSCC
and the vehicle were on the same page after an install. Usually
validation after install required the joystick commander application
but that provides no indication as to whether the vehicle reports
what we expect it to after sending commands.

After this commit the OSCC repo will have the `consensus.py` program
to verify tat OSCC and the vehicle are on the same page or report
otherwise. Ideally this program will be the first step taken after
installing OSCC. It will also help pinpoint issues if the vehicle
provides unexpected or error results if OSCC expected different
information.