Update the CMake include path
Prior to this commit the CMake API include used a variable assuming the
sub-directory would be within the api which does not work if the oscc
repo is used as a submodule. This commit fixes that by using a current
directory variable instead so that the include file still works as a
submodule.